refactor: 편지 레거시 코드 삭제
