+++ src/main/java/com/handwoong/rainbowletter/common/config/chatgpt/ChatGptConfig.java
@@ -1,4 +1,4 @@
-package com.handwoong.rainbowletter.common.config;
+package com.handwoong.rainbowletter.common.config.chatgpt;
 
 import lombok.Getter;
 import org.springframework.beans.factory.annotation.Value;
@@ -6,10 +6,7 @@ import org.springframework.context.annotation.Configuration;
 
 @Getter
 @Configuration
-public class PropertiesConfig {
+public class ChatGptConfig {
     @Value("${chatgpt.token}")
-    private String chatgptToken;
-
-    @Value("${airtable.token}")
-    private String airtableToken;
+    private String chatGptToken;
 }
