--- src/main/java/com/handwoong/rainbowletter/letter/controller/port/LetterLegacyService.java
@@ -1,7 +1,0 @@
-package com.handwoong.rainbowletter.letter.controller.port;
-
-import com.handwoong.rainbowletter.letter.dto.ReplyRequestDto;
-
-public interface LetterLegacyService {
-    void reply(final ReplyRequestDto replyRequest);
-}
