--- src/main/java/com/handwoong/rainbowletter/letter/dto/AirtableUpdateDto.java
@@ -1,9 +1,0 @@
-package com.handwoong.rainbowletter.letter.dto;
-
-public record AirtableUpdateDto(
-        String GPT,
-        int prompt_tokens,
-        int completion_tokens,
-        int total_tokens
-) {
-}
