--- src/main/java/com/handwoong/rainbowletter/letter/dto/AirtableUpdateRequestDto.java
@@ -1,6 +1,0 @@
-package com.handwoong.rainbowletter.letter.dto;
-
-public record AirtableUpdateRequestDto(
-        AirtableUpdateDto fields
-) {
-}
