--- src/main/java/com/handwoong/rainbowletter/letter/dto/ChatGptMessage.java
@@ -1,7 +1,0 @@
-package com.handwoong.rainbowletter.letter.dto;
-
-public record ChatGptMessage(
-        String role,
-        String content
-) {
-}
