--- src/main/java/com/handwoong/rainbowletter/letter/dto/ChatGptRequestDto.java
@@ -1,14 +1,0 @@
-package com.handwoong.rainbowletter.letter.dto;
-
-import java.util.List;
-
-public record ChatGptRequestDto(
-        String model,
-        List<ChatGptMessage> messages,
-        Long max_tokens,
-        Double temperature,
-        Double top_p,
-        Double frequency_penalty,
-        Double presence_penalty
-) {
-}
