--- src/main/java/com/handwoong/rainbowletter/letter/dto/ChatGptResponseChoice.java
@@ -1,8 +1,0 @@
-package com.handwoong.rainbowletter.letter.dto;
-
-public record ChatGptResponseChoice(
-        ChatGptMessage message,
-        int index,
-        String finish_reason
-) {
-}
