--- src/main/java/com/handwoong/rainbowletter/letter/dto/ChatGptResponseUsage.java
@@ -1,8 +1,0 @@
-package com.handwoong.rainbowletter.letter.dto;
-
-public record ChatGptResponseUsage(
-        int prompt_tokens,
-        int completion_tokens,
-        int total_tokens
-) {
-}
