--- src/main/java/com/handwoong/rainbowletter/letter/dto/ReplyRequestDto.java
@@ -1,9 +1,0 @@
-package com.handwoong.rainbowletter.letter.dto;
-
-public record ReplyRequestDto(
-        String baseId,
-        String recordId,
-        String tableName,
-        ChatGptRequestDto body
-) {
-}
